crypto_risk_analytics/requirements.txt
pandas
numpy
scipy
ccxt
influxdb-client
requests
python-dotenv
schedule
scikit-learn
nltk
beautifulsoup4
